feat: add curve spell

+++ contracts/wrapper/WCurveGauge.sol
@@ -37,7 +37,7 @@ contract WCurveGauge is
 
     struct GaugeInfo {
         ILiquidityGauge impl; // Gauge implementation
-        uint accCrvPerShare; // Accumulated CRV per share
+        uint256 accCrvPerShare; // Accumulated CRV per share
     }
 
     /// @dev Address of Curve Registry
@@ -45,7 +45,7 @@ contract WCurveGauge is
     /// @dev Address of CRV token
     IERC20Upgradeable public CRV;
     /// @dev Mapping from pool id to (mapping from gauge id to GaugeInfo)
-    mapping(uint => mapping(uint => GaugeInfo)) public gauges;
+    mapping(uint256 => mapping(uint256 => GaugeInfo)) public gauges;
 
     function initialize(
         address crv_,
@@ -57,49 +57,49 @@ contract WCurveGauge is
         registry = ICurveRegistry(crvRegistry_);
     }
 
-    /// @dev Encode pid, gid, crvPerShare to a ERC1155 token id
+    /// @notice Encode pid, gid, crvPerShare to a ERC1155 token id
     /// @param pid Curve pool id (10-bit)
     /// @param gid Curve gauge id (6-bit)
     /// @param crvPerShare CRV amount per share, multiplied by 1e18 (240-bit)
     function encodeId(
-        uint pid,
-        uint gid,
-        uint crvPerShare
-    ) public pure returns (uint) {
+        uint256 pid,
+        uint256 gid,
+        uint256 crvPerShare
+    ) public pure returns (uint256) {
         require(pid < (1 << 10), "bad pid");
         require(gid < (1 << 6), "bad gid");
         require(crvPerShare < (1 << 240), "bad crv per share");
         return (pid << 246) | (gid << 240) | crvPerShare;
     }
 
-    /// @dev Decode ERC1155 token id to pid, gid, crvPerShare
+    /// @notice Decode ERC1155 token id to pid, gid, crvPerShare
     /// @param id Token id to decode
     function decodeId(
-        uint id
-    ) public pure returns (uint pid, uint gid, uint crvPerShare) {
+        uint256 id
+    ) public pure returns (uint256 pid, uint256 gid, uint256 crvPerShare) {
         pid = id >> 246; // First 10 bits
         gid = (id >> 240) & (63); // Next 6 bits
         crvPerShare = id & ((1 << 240) - 1); // Last 240 bits
     }
 
-    /// @dev Get underlying ERC20 token of ERC1155 given pid, gid
+    /// @notice Get underlying ERC20 token of ERC1155 given pid, gid
     /// @param pid pool id
     /// @param gid gauge id
     function getUnderlyingTokenFromIds(
-        uint pid,
-        uint gid
+        uint256 pid,
+        uint256 gid
     ) public view returns (address) {
         ILiquidityGauge impl = gauges[pid][gid].impl;
         if (address(impl) == address(0)) revert Errors.NO_GAUGE();
         return impl.lp_token();
     }
 
-    /// @dev Get underlying ERC20 token of ERC1155 given token id
+    /// @notice Get underlying ERC20 token of ERC1155 given token id
     /// @param id Token id
     function getUnderlyingToken(
-        uint id
+        uint256 id
     ) external view override returns (address) {
-        (uint pid, uint gid, ) = decodeId(id);
+        (uint256 pid, uint256 gid, ) = decodeId(id);
         return getUnderlyingTokenFromIds(pid, gid);
     }
 
@@ -109,7 +109,7 @@ contract WCurveGauge is
     /// @param amount amount of share
     function pendingRewards(
         uint256 tokenId,
-        uint amount
+        uint256 amount
     )
         public
         view
@@ -119,10 +119,10 @@ contract WCurveGauge is
         // TODO:
     }
 
-    /// @dev Register curve gauge to storage given pool id and gauge id
+    /// @notice Register curve gauge to storage given pool id and gauge id
     /// @param pid Pool id
     /// @param gid Gauge id
-    function registerGauge(uint pid, uint gid) external onlyOwner {
+    function registerGauge(uint256 pid, uint256 gid) external onlyOwner {
         if (address(gauges[pid][gid].impl) != address(0))
             revert Errors.EXISTING_GAUGE(pid, gid);
         address pool = registry.pool_list(pid);
@@ -141,15 +141,15 @@ contract WCurveGauge is
         });
     }
 
-    /// @dev Mint ERC1155 token for the given ERC20 token
+    /// @notice Mint ERC1155 token for the given ERC20 token
     /// @param pid Pool id
     /// @param gid Gauge id
     /// @param amount Token amount to wrap
     function mint(
-        uint pid,
-        uint gid,
-        uint amount
-    ) external nonReentrant returns (uint) {
+        uint256 pid,
+        uint256 gid,
+        uint256 amount
+    ) external nonReentrant returns (uint256) {
         GaugeInfo storage gauge = gauges[pid][gid];
         ILiquidityGauge impl = gauge.impl;
         if (address(impl) == address(0)) revert Errors.NO_GAUGE();
@@ -157,19 +157,22 @@ contract WCurveGauge is
         IERC20Upgradeable lpToken = IERC20Upgradeable(impl.lp_token());
         lpToken.safeTransferFrom(msg.sender, address(this), amount);
         impl.deposit(amount);
-        uint id = encodeId(pid, gid, gauge.accCrvPerShare);
+        uint256 id = encodeId(pid, gid, gauge.accCrvPerShare);
         _mint(msg.sender, id, amount, "");
         return id;
     }
 
-    /// @dev Burn ERC1155 token to redeem ERC20 token back
+    /// @notice Burn ERC1155 token to redeem ERC20 token back
     /// @param id Token id to burn
     /// @param amount Token amount to burn
-    function burn(uint id, uint amount) external nonReentrant returns (uint) {
+    function burn(
+        uint256 id,
+        uint256 amount
+    ) external nonReentrant returns (uint256) {
         if (amount == type(uint256).max) {
             amount = balanceOf(msg.sender, id);
         }
-        (uint pid, uint gid, uint stCrvPerShare) = decodeId(id);
+        (uint256 pid, uint256 gid, uint256 stCrvPerShare) = decodeId(id);
         _burn(msg.sender, id, amount);
         GaugeInfo storage gauge = gauges[pid][gid];
         ILiquidityGauge impl = gauge.impl;
@@ -177,23 +180,23 @@ contract WCurveGauge is
         _mintCrv(gauge);
         impl.withdraw(amount);
         IERC20Upgradeable(impl.lp_token()).safeTransfer(msg.sender, amount);
-        uint stCrv = (stCrvPerShare * amount) / 1e18;
-        uint enCrv = (gauge.accCrvPerShare * amount) / 1e18;
+        uint256 stCrv = (stCrvPerShare * amount) / 1e18;
+        uint256 enCrv = (gauge.accCrvPerShare * amount) / 1e18;
         if (enCrv > stCrv) {
             CRV.safeTransfer(msg.sender, enCrv - stCrv);
         }
         return pid;
     }
 
-    /// @dev Mint CRV reward for curve gauge
+    /// @notice Mint CRV reward for curve gauge
     /// @param gauge Curve gauge to mint reward
     function _mintCrv(GaugeInfo storage gauge) internal {
         ILiquidityGauge impl = gauge.impl;
-        uint balanceBefore = CRV.balanceOf(address(this));
+        uint256 balanceBefore = CRV.balanceOf(address(this));
         ILiquidityGaugeMinter(impl.minter()).mint(address(impl));
-        uint balanceAfter = CRV.balanceOf(address(this));
-        uint gain = balanceAfter - balanceBefore;
-        uint supply = impl.balanceOf(address(this));
+        uint256 balanceAfter = CRV.balanceOf(address(this));
+        uint256 gain = balanceAfter - balanceBefore;
+        uint256 supply = impl.balanceOf(address(this));
         if (gain > 0 && supply > 0) {
             gauge.accCrvPerShare += (gain * 1e18) / supply;
         }
